pyproject.toml
[project]
name = "axiom-2-0"
version = "0.1.0"
description = "An Advanced version of Axiom Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.11",
    "prompt-toolkit>=3.0.51",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/axiom"] 

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
